requirements.txt
absl-py==1.2.0
aiohttp==3.8.3
aiosignal==1.2.0
alembic==1.8.1
aniso8601==9.0.1
anyascii==0.3.1
appdirs==1.4.4
argcomplete==2.0.0
arrow==1.2.3
asttokens==2.0.8
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
audioread==3.0.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
blinker==1.5
blis==0.7.8
cachetools==5.2.0
cardboardlint==1.3.0
catalogue==2.0.8
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
ChatterBot==1.0.4
chatterbot-corpus==1.2.0
ci-sdr==0.0.2
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
comtypes==1.1.14
confection==0.0.2
ConfigArgParse==1.5.3
contourpy==1.0.5
coqpit==0.0.16
cryptography==38.0.1
ctc-segmentation==1.7.1
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
dateparser==1.1.1
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
Distance==0.1.3
distlib==0.3.6
docopt==0.6.2
einops==0.4.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
-e git+https://github.com/SAOCyper/GLADOS.git@5cb067751981b0b941c3846b5a8fab81e5c49732#egg=espnet
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
executing==1.0.0
fast-bss-eval==0.1.3
ffmpeg==1.4
filelock==3.8.0
Flask==2.2.2
Flask-Admin==1.6.0
Flask-Assets==2.0
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-Ext==0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Markdown==0.3
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==2.0.7
fonttools==4.37.2
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
g2p-en==2.1.0
g2pkk==0.1.2
gast==0.4.0
gdown==4.5.1
google-auth==2.11.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.3
grpcio==1.48.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-cs==2.0.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
gruut-lang-it==2.0.0
gruut-lang-nl==2.0.2
gruut-lang-pt==2.0.0
gruut-lang-ru==2.0.0
gruut-lang-sv==2.0.0
gunicorn==20.1.0
h11==0.13.0
h5py==3.7.0
huggingface-hub==0.9.1
humanfriendly==10.0
ics==0.7.2
idna==3.4
importlib-metadata==4.12.0
inflect==5.6.0
ipykernel==6.15.3
ipython==8.5.0
iso8601==1.0.2
itsdangerous==2.1.2
jaconv==0.3
jamo==0.4.1
jedi==0.18.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
jsonlines==1.2.0
jupyter-core==4.11.1
jupyter_client==7.3.5
kaldiio==2.17.2
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
libclang==14.0.6
librosa==0.8.0
llvmlite==0.38.1
Mako==1.2.3
Markdown==3.4.1
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib==3.6.0
matplotlib-inline==0.1.6
mecab-python3==1.0.5
mod-wsgi==4.9.4
multidict==6.0.2
murmurhash==1.0.8
nest-asyncio==1.5.5
networkx==2.8.6
nltk==3.7
nose==1.3.7
num2words==0.5.12
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
outcome==1.2.0
packaging==21.3
pandas==1.4.4
parallel-wavegan==0.5.5
parso==0.8.3
passlib==1.7.4
pathy==0.10.1
pdfminer.six==20220524
pdfplumber==0.7.4
peewee==3.15.3
pickleshare==0.7.5
Pillow==9.2.0
Pint==0.20.1
pipreqs==0.4.11
platformdirs==2.5.2
playsound==1.3.0
pooch==1.6.0
preshed==3.0.7
prompt-toolkit==3.0.31
protobuf==3.19.5
psutil==5.9.2
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.12
pycparser==2.21
pydantic==1.9.2
pydub==0.25.1
Pygments==2.13.0
pylint==2.5.3
pymongo==3.13.0
pynndescent==0.5.7
pyopenjtalk==0.2.0
pyparsing==3.0.9
pypinyin==0.44.0
pypiwin32==223
pyreadline3==3.4.1
pysbd==0.3.4
pyserial==3.5
PySocks==1.7.1
python-crfsuite==0.9.8
python-dateutil==2.7.5
python-dotenv==0.21.0
pytorch-wpe==0.0.1
pyttsx==1.1
pyttsx3==2.90
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pywin32==228
pyworld==0.2.10
PyYAML==3.13
pyzmq==24.0.0
redis==4.3.4
regex==2022.3.2
replit==3.2.4
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rich==12.5.1
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.1
selenium==4.4.3
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.3.2.post1
sox==1.4.1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.3
speaklater==1.3
SpeechRecognition==3.8.1
spotipy==2.20.0
SQLAlchemy==1.2.19
srsly==2.4.4
stack-data==0.5.0
TatSu==5.8.3
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.5.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
thinc==8.1.2
threadpoolctl==3.1.0
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0+cu113
torch-complex==0.4.3
torchaudio==0.11.0+cu113
torchvision==0.12.0+cu113
tornado==6.2
tqdm==4.64.1
trainer==0.0.15
traitlets==5.4.0
transformers==4.22.2
trio==0.21.0
trio-websocket==0.9.2
-e git+https://github.com/coqui-ai/TTS@d6ad9a05b40e9f3aa876ecfa1c0d23339f373df4#egg=TTS
typeguard==2.13.3
typer==0.4.2
typing-extensions==3.10.0.2
tzdata==2022.2
tzlocal==4.2
umap-learn==0.5.1
Unidecode==1.3.4
unidic-lite==1.0.8
urllib3==1.26.12
utility==1.0
virtualenv==20.16.5
waitress==2.1.2
Wand==0.6.10
wasabi==0.10.1
wcwidth==0.2.5
webassets==2.0
Werkzeug==2.2.2
wfastcgi==3.0.0
wget==3.2
wikipedia==1.4.0
wrapt==1.14.1
wsproto==1.2.0
WTForms==3.0.1
xmltodict==0.13.0
yarg==0.1.9
yarl==1.8.1
yq==3.1.0
zipp==3.8.1
